// server/src/types/game.types.ts
export interface IGameUser {
    _id: string;
    username: string;
    avatar: string;
    balance: number;
}

export interface IPlayer {
    socketId: string;
    user: IGameUser;
}

export interface IBot {
    socketId: string;
    user: {
        _id: string;
        username: string;
        avatar: string;
        balance: number;
    };
}

export type GamePlayer = IPlayer | IBot;
// ...
export interface ITicTacToeState extends IBaseGameState {
    board: ('X' | 'O' | null)[];
}

export interface IChessState extends IBaseGameState {
    fen: string;
    gameHistory: string[];
}

export interface ICheckersState extends IBaseGameState {
    board: (IPiece | null)[];
    mustCaptureWith: number | null;
}

export interface IBackgammonState extends IBaseGameState {
    points: IPoint[];
    dice: number[];
    turnPhase: 'ROLLING' | 'MOVING';
    borneOff: [number, number];
}

export interface IPiece {
    playerIndex: 0 | 1;
    isKing: boolean;
}

export type IPoint = [playerIndex: 0 | 1, count: number] | null;

// Типы для ходов
export interface IBasicMove {
    from?: number | string;
    to?: number | string;
}

export interface ITicTacToeMove {
    cellIndex: number;
}

export interface IChessMove {
    from: string;
    to: string;
    promotion?: string;
}

export interface ICheckersMove extends IBasicMove {
    from: number;
    to: number;
    isCapture: boolean;
}

export interface IBackgammonMove extends IBasicMove {
    from: number;
    to: number;
}

export type GameMove = ITicTacToeMove | IChessMove | ICheckersMove | IBackgammonMove | Record<string, any>;
export type GameState = ITicTacToeState | IChessState | ICheckersState | IBackgammonState | Record<string, any>;

// Типы для комнат
export interface IRoom {
    id: string;
    gameType: 'tic-tac-toe' | 'checkers' | 'chess' | 'backgammon';
    bet: number;
    players: GamePlayer[];
    gameState: GameState;
    botJoinTimer?: NodeJS.Timeout;
    disconnectTimer?: NodeJS.Timeout;
}